+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - Fix compatibility with Scavenger plugin
  - Add "use_magic_damage" config option to turn off "magic" (splash potion) damage source
  - Fix XP scale to 1.8 standards (effects XP-based shops)
+ - Add "regenerate_while_inactive" config option to turn off wands regen'ing mana while not held
 
 ## 4.9.5.1